[ingridDatabase] dispatch to Tpx1/3 ToT file reading proc, pulses floats

+++ Analysis/ingrid/ingrid_types.nim
@@ -68,7 +68,7 @@ type
   #############################
 
   Tot* = object
-    pulses*: seq[int]
+    pulses*: seq[float]
     mean*: seq[float]
     std*: seq[float]
